Fix CardHeader props typing and collapsible state sync
